requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
motor==3.3.2
scikit-learn==1.3.2
